+++ apps/backend/src/modules/friends/friends.ts
@@ -9,7 +9,8 @@ const friendRepository = new FriendRepository(UserModel);
 const friendService = new FriendService(friendRepository);
 const friendController = new FriendController(logger, friendService);
 
-export { friendController };
+export { friendController, friendRepository };
+export { FriendRepository } from "./friend.repository.js";
 export {
 	addFriendValidationSchema,
 	friendIdParameterValidationSchema,