+++ packages/shared/src/modules/chats/chats.ts
@@ -1,0 +1,12 @@
+export { ChatsApiPath } from "./libs/enums/enums.js";
+export { ChatError } from "./libs/exceptions/exceptions.js";
+export {
+	type ChatCreateRequestDto,
+	type ChatGetAllItemResponseDto,
+	type ChatItemResponseDto,
+	type ChatResponseDto,
+} from "./libs/types/types.js";
+export {
+	chatIdParameter as chatIdParameterValidationSchema,
+	chatMessageCreate as chatCreateValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";