+++ packages/shared/src/modules/chats/libs/enums/chat-validation-rule.enum.ts
@@ -1,0 +1,6 @@
+const ChatValidationRule = {
+	CHAT_ID_MINIMUM_VALUE: 1,
+	USER_ID_MINIMUM_VALUE: 1,
+} as const;
+
+export { ChatValidationRule };